[update] integration / 他のintegration testでも使う共通関数・変数のファイル分け
